Factor SessionManager's process launching out to PowerShellProcess
This change moves SessionManager's process launching behavior into a new
class called PowerShellProcess which will be used for launching
additional PowerShell processes to enable fresh, transient debugging
sessions.
